+++ tests/test_test_postgres.py
@@ -68,6 +68,21 @@ def test_test_postgres_case_sensitive_table_name(postgres_container, monkeypatch
     assert all(check.result == "passed" for check in run.checks)
 
 
+def test_test_postgres_servicelevels_freshness_should_fail(postgres_container, monkeypatch):
+    monkeypatch.setenv("DATACONTRACT_POSTGRES_USERNAME", postgres.username)
+    monkeypatch.setenv("DATACONTRACT_POSTGRES_PASSWORD", postgres.password)
+    _init_sql("fixtures/postgres/data/data.sql")
+
+    datacontract_file = "fixtures/postgres/datacontract_servicelevels.yaml"
+    data_contract_str = _setup_datacontract(datacontract_file)
+    data_contract = DataContract(data_contract_str=data_contract_str)
+
+    run = data_contract.test()
+
+    print(run.pretty())
+    assert run.result == "failed"
+
+
 def _setup_datacontract(file):
     with open(file) as data_contract_file:
         data_contract_str = data_contract_file.read()