+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `datacontract test` now also executes tests for service levels freshness and retention (#407)
+
 ### Changed
 
 ### Fixed